cloud_agnostic_agent/requirements.txt
langgraph
langchain
openai
requests
google-cloud-secret-manager
google-auth
langchain-google-genai
google-generativeai